requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
bleach==6.1.0
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.8.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
cryptography==44.0.3
cssselect2==0.8.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-apscheduler==0.7.0
django-crontab==0.7.1
djangorestframework==3.16.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.104.1
ffmpeg-python==0.2.0
fpdf==1.7.2
frozenlist==1.6.0
future==1.0.0
greenlet==3.2.2
gunicorn==21.2.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.2
jose==1.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.3
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.4.3
mutagen==1.47.0
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
parso==0.8.4
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.48
propcache==0.3.1
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.0
pypng==0.20220715.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.6
pytube==15.0.0
pywin32==308
pyzmq==26.2.0
qrcode==7.4.2
redis==6.1.0
reportlab==4.2.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.27.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
twitchio==2.8.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.24.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarl==1.20.0
yt-dlp==2024.3.10
